% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{causalEffectsSimulation}
\alias{causalEffectsSimulation}
\title{Simulate BAPs and compute causal effects of greedy search results and ground truth}
\usage{
causalEffectsSimulation(N, p, n, mc.cores = 1, max.in.degree = Inf,
  n.restarts = 0, max.steps = Inf, max.iter.ricf = 10,
  equivalent.eps = 1e-10, verbose = FALSE)
}
\arguments{
\item{N}{Number of simulation runs}

\item{p}{Number of vertices}

\item{n}{Number of samples to generate}

\item{mc.cores}{Parallelism (number of cores to be used). If set to 1, no parallelism is used. Should be at most N.}

\item{max.in.degree}{Maximal allowed in-degree}

\item{n.restarts}{Number of restarts}

\item{max.steps}{Max number of greedy steps per run}

\item{max.iter.ricf}{Max number of iterations in RICF}

\item{equivalent.eps}{Maximal allowed score difference for two graphs to be considered equivalent}

\item{verbose}{Print extra log output?}
}
\value{
List containing:
  CE.minabs.gt - List of minimal absolute causal effect matrices E (one per simulation run)
    for the ground truths, where (E)_i,j is the causal effect on j on i
  CE.minabs.greedy - same for the greedy search results
}
\description{
Simulate BAPs and compute causal effects of greedy search results and ground truth
}
