% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{generateBAPPArameters}
\alias{generateBAPPArameters}
\title{Randomly generate parameters (edge weights) for a given BAP}
\usage{
generateBAPPArameters(bap)
}
\arguments{
\item{bap}{p-by-p adjacency matrix}
}
\value{
List containing:
  B - p-by-p matrix of directed edge weights
  Omega - p-by-p matrix of bidirected edge weights
}
\description{
We consider the following linear recursive model:
  X = B^t X + eps,
where (B)_{i,j} is the edge weight of edge i->j and corresponds to the (i,j)-entry
in the adjacency matrix of the graph, and eps has a N(0, Omega) distribution.
}
