% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{generateData}
\alias{generateData}
\title{Randomly generate data for a given BAP parametrisation}
\usage{
generateData(n, params)
}
\arguments{
\item{n}{Number of samples to generate}

\item{params}{Named list containing:
B - p-by-p matrix of directed edge weights
Omega - p-by-p matrix of bidirected edge weights}
}
\value{
n-by-p matrix
}
\description{
Randomly generate data for a given BAP parametrisation
}
