% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{generateUniformBAP}
\alias{generateUniformBAP}
\title{Uniform BAP generation using MCMC}
\usage{
generateUniformBAP(p, N = 1, dag = FALSE, max.in.degree = Inf,
  names = paste0("V", 1:p), iter = 6 * p^2)
}
\arguments{
\item{p}{Number of vertices}

\item{N}{Number of graphs to generate}

\item{dag}{Whether to generate DAGs or more general BAPs}

\item{max.in.degree}{Maximal allowed in-degree}

\item{names}{Names to give to the rows / columns of the adjacency matrix}

\item{iter}{Number if MCMC iterations between samples}
}
\value{
List of N p-by-p adjacency matrices
}
\description{
Uniform BAP generation using MCMC
}
