% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{greedySearch}
\alias{greedySearch}
\title{Greedy BAP search}
\usage{
greedySearch(data, mg.start = NULL, n.restarts = 0, mc.cores = 1,
  max.steps = Inf, max.iter.ricf = 10, neighbourhood.size = Inf,
  eps.conv = 1e-12, max.in.degree = Inf, dags.only = FALSE,
  direction = "both", verbose = FALSE)
}
\arguments{
\item{data}{n-by-p data matrix (n samples, p variables)}

\item{mg.start}{Starting graph(s). Must be either NULL, an adjacency matrix defining the mixed graph (if n.restarts is 0) or a list of adjacency matrices of length n.restarts+1. If NULL, a starting graph is uniformly sampled for each run.}

\item{n.restarts}{Number of restarts}

\item{mc.cores}{Parallelism (number of cores to be used). If set to 1, no parallelism is used. Should be at most n.restarts+1.}

\item{max.steps}{Max number of greedy steps per run}

\item{max.iter.ricf}{Max number of iterations in RICF}

\item{neighbourhood.size}{If set to a finite number, only this many (randomly sampled) neighbours are considered at each greedy step}

\item{eps.conv}{Minimal score improvement to continue search after each step}

\item{max.in.degree}{Max in-degree of considered graphs}

\item{dags.only}{Consider DAGs only?}

\item{direction}{Can be "forward" (only adding or changing edges), "backward" (only removing or changing edges), or "both" (adding, removing, and changing edges)}

\item{verbose}{Print extra log output?}
}
\value{
List of n.restarts+1 greedy results, where each greedy result is a list containing:
}
\description{
Greedy BAP search
}
