% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBAP}
\alias{plotBAP}
\title{Plots a BAP using circular layout}
\usage{
plotBAP(bap, tcltk = FALSE, ...)
}
\arguments{
\item{bap}{p-by-p adjacency matrix}

\item{...}{any extra arguments to pass on to ggm::plotGraph}
}
\description{
Plots a BAP using circular layout
}
