% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotROCCurve}
\alias{plotROCCurve}
\title{Plots ROC curve from estimated and true causal effects}
\usage{
plotROCCurve(CE.gt, CE.greedy, tol = 1e-08, precision.recall = FALSE,
  avg.auc = TRUE)
}
\arguments{
\item{CE.gt}{List of minimal absolute causal effect matrices E (one per simulation run)
for the ground truths, where (E)_i,j is the causal effect on j on i (as returned by
causalEffectsSimulation)}

\item{CE.greedy}{same for the greedy search results}

\item{tol}{Tolerance for causal effects (everythin smaller than this will be thresholded)}

\item{precision.recall}{Plot precision-recall instead of FPR / TPR?}

\item{avg.auc}{If TRUE: Average of AUCs of individual curves
If FALSE: AUC of averaged curves}
}
\value{
List of three elements:
  FPR - Matrix with false positive rates for each of the simulation runs (one column per run)
  TPR - Matrix with true positive rates for each of the simulation runs (one column per run)
  auc - Averaged AUC
}
\description{
Plots ROC curve from estimated and true causal effects
}
