% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotScoreCurves}
\alias{plotScoreCurves}
\title{Plots scores per greedy step against cumulative time}
\usage{
plotScoreCurves(scores, times, plot.steps = FALSE, ylim = NULL,
  title.start = "Max Score: ")
}
\arguments{
\item{scores}{List (one element per greedy search) with score vectors (as returned by greedySearch)}

\item{times}{List (one element per greedy search) with cumulative time vectors (as returned by greedySearch)}

\item{plot.steps}{Should a second plot of scores vs steps be generated?}

\item{ylim}{Vector (ymin, ymax) - if this is NULL, it is generated from scores}

\item{title.start}{String with title prefix (the max score is appended to this)}
}
\description{
Plots scores per greedy step against cumulative time
}
